First attempt at logstash
This is a first attempt to add logstash. It does so by connecting
machines in the topology:

 * node1.internal rsyslog daemon, transports to
 * logging.internal rsyslog daemon, transports to
 * logging.internal logstash daemon

However there is something funny in the config where local messages on
logging.internal are transported to logstash, and remote messages from
node1.internal are saved to disk by logging.internal rsyslog, but
node1.internal messages aren't then being delivered to logstash.

This commit is mainly to save the work that I had.